% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgarfundamentals-package.R
\docType{package}
\name{edgarfundamentals-package}
\alias{edgarfundamentals}
\alias{edgarfundamentals-package}
\title{edgarfundamentals: Retrieve Fundamental Financial Data from SEC EDGAR}
\description{
Provides a simple, ticker-based interface for retrieving fundamental financial data from the United States Securities and Exchange Commission's EDGAR XBRL API \url{https://www.sec.gov/edgar/sec-api-documentation}. Functions return key financial ratios including earnings per share, return on equity, return on assets, debt-to-equity, current ratio, gross margin, operating margin, net margin, price-to-earnings, price-to-book, and dividend yield for any publicly traded U.S. company. Data is sourced directly from company 10-K annual filings, requiring no API key or paid subscription. Designed for use in quantitative finance courses and research workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robschumaker/edgarfundamentals}
  \item Report bugs at \url{https://github.com/robschumaker/edgarfundamentals/issues}
}

}
\author{
\strong{Maintainer}: Robert Schumaker \email{rob.schumaker@gmail.com} (\href{https://orcid.org/0000-0003-2977-1856}{ORCID})

}
\keyword{internal}
